Rearrange implementation files
