+++ heart/heart-core/src/fibers/system.cpp
@@ -17,7 +17,7 @@
 
 #include "heart/debug/assert.h"
 
-#include "fibers/abi_context.h"
+#include "fibers/native_impl.h"
 
 #include <stdio.h>
 
@@ -53,8 +53,7 @@ void* HeartFiberSystem::HeartFiberThreadEntry(void* parameter)
 	currentThread = nullptr;
 
 	// Allocate a startup abi so that the system has something to jump away from
-	auto* startupAbi = HeartFiberContext::Get().currentSystem->m_allocator.AllocateAndConstruct<HeartFiberAbiContext>();
-	HeartFiberContext::Get().entryUnit.m_nativeHandle = startupAbi;
+	HeartFiberContext::Get().currentSystem->InitializeEntryFiber();
 	HeartFiberContext::Get().currentWorkUnit = &HeartFiberContext::Get().entryUnit;
 
 	// Initialize and execute the pump
@@ -64,12 +63,14 @@ void* HeartFiberSystem::HeartFiberThreadEntry(void* parameter)
 	// We've returned from the pump, time to cleanup and kill the thread.
 	HEART_ASSERT(HeartFiberContext::Get().currentSystem->m_exit == true);
 	HeartFiberContext::Get().currentSystem->ReleaseWorkUnitNativeHandle(HeartFiberContext::Get().pump);
-	HeartFiberContext::Get().currentSystem->ReleaseWorkUnitNativeHandle(HeartFiberContext::Get().entryUnit);
+
+	// Release the entry fiber - actual behavior is implementation-dependent (since we're currently *on* it).
+	HeartFiberContext::Get().currentSystem->ReleaseEntryFiber();
 
 	return nullptr;
 }
 
-void HeartFiberSystem::HeartFiberStartRoutine()
+void HeartFiberSystem::HeartFiberStartRoutine(void*)
 {
 	// Verify that the fiber native code gave us a valid stack
 	HEART_FIBER_VERIFY_STACK();
@@ -150,68 +151,6 @@ HeartFiberStatus HeartFiberSystem::PumpRoutine()
 	return HeartFiberStatus::Complete;
 }
 
-void HeartFiberSystem::InitializeWorkUnitNativeHandle(HeartFiberWorkUnit& unit)
-{
-	// We can only create fibers from a fiber!
-	HEART_ASSERT(HeartFiberContext::Get().currentSystem != nullptr);
-
-	// TODO: Large stack and small stack fibers
-	// SFML requires a fairly deep stack (at least 256kb) so default to large for now
-	constexpr uint32_t StackSize = 1u * Meg;
-	byte_t* stackStart = m_allocator.allocate<byte_t>(StackSize);
-
-	HeartFiberAbiContext* ctx = m_allocator.AllocateAndConstruct<HeartFiberAbiContext>();
-	ctx->allocated = stackStart;
-
-	// Stacks grow downward, point to the end of our buffer, and "allocate"
-	// required Windows "red zone" space of 4 64bit values
-	byte_t* stackEnd = stackStart + StackSize;
-	byte_t* rsp = stackEnd - (4 * sizeof(uint64_t));
-	if ((uintptr_t(rsp) & 8L) != 8L)
-	{
-		// XXX: Windows requires that the stack is always 16-byte aligned... except
-		// during the PROLOG of a function, during which it expects it to be "misaligned"
-		// by 8 due to the push of the return pointer. It thus always emits code that
-		// offsets rsp by 8 + (x * 16). The end result is that if our stack pointer starts
-		// off properly aligned, the PROLOG of the StartRoutine misaligns it and causes
-		// horrible crashes later. So, we need to purposefully ensure it is *not* 16-byte
-		// aligned to start with.
-		rsp -= 8;
-	}
-
-#if !HEART_STRICT_PERF
-	// Zero the whole buffer if we're not in strict perf
-	memset(stackStart, 0, StackSize);
-#else
-	// Zero just the area between our stack pointer and the end if we're in release
-	// This isn't strictly necessary, but debugging is already hard enough in
-	// optimized builds, we don't need a nonsense garbage stack to contend with too.
-	memset(rsp, 0, stackEnd - rsp);
-#endif
-
-	ctx->rsp = uintptr_t(rsp);
-	ctx->rbp = uintptr_t(rsp);
-	ctx->rip = uintptr_t(&HeartFiberStartRoutine);
-
-	unit.m_nativeHandle = ctx;
-}
-
-void HeartFiberSystem::ReleaseWorkUnitNativeHandle(HeartFiberWorkUnit& unit)
-{
-	HEART_ASSERT(HeartFiberContext::Get().currentSystem != nullptr);
-
-	HeartFiberAbiContext* ctx = reinterpret_cast<HeartFiberAbiContext*>(unit.m_nativeHandle);
-	unit.m_nativeHandle = nullptr;
-
-	// Free the stack
-	m_allocator.deallocate(ctx->allocated);
-	ctx->allocated = nullptr;
-
-	// Free the actual abi context
-	m_allocator.DestroyAndFree<HeartFiberAbiContext>(ctx);
-	ctx = nullptr;
-}
-
 void HeartFiberSystem::CompleteWorkUnit(HeartFiberWorkUnit& unit)
 {
 	HeartFiberContext::Get().destroyQueue.PushBack(&unit);
@@ -240,21 +179,6 @@ void HeartFiberSystem::YieldUnit(HeartFiberWorkUnit& unit)
 
 	// And verify upon coming back that we have a valid stack
 	HEART_FIBER_VERIFY_STACK();
-}
-
-void HeartFiberSystem::NativeSwitchToFiber(HeartFiberWorkUnit& target)
-{
-	// Get where we're coming from
-	HeartFiberWorkUnit* from = HeartFiberContext::Get().currentWorkUnit;
-	HeartFiberAbiContext* from_abi = reinterpret_cast<HeartFiberAbiContext*>(from->m_nativeHandle);
-
-	// Populate the context
-	HEART_ASSERT(HeartFiberContext::Get().currentSystem == this);
-	HeartFiberContext::Get().currentWorkUnit = &target;
-
-	// Execute
-	HeartFiberAbiContext* to_abi = reinterpret_cast<HeartFiberAbiContext*>(target.m_nativeHandle);
-	heart_swap_fiber_context(from_abi, to_abi);
 }
 
 void HeartFiberSystem::NativeSwitchToFiberNoReturn(HeartFiberWorkUnit& target)
