+++ heart/heart-test/src/fibers.cpp
@@ -184,46 +184,3 @@ TEST(Fibers, FourThreads)
 
 	system.Shutdown();
 }
-
-TEST(Fibers, Allocation)
-{
-	TestTrackingAllocator allocator;
-
-	{
-		HeartFiberSystem::Settings settings;
-		settings.threadCount = 4;
-
-		HeartFiberSystem system(allocator);
-		system.Initialize(settings);
-
-		const int JobCount = 128;
-		std::atomic_int doneCount = 0;
-
-		std::mutex threadIdsMutex;
-		std::vector<std::thread::id> threadIds;
-
-		int i = JobCount;
-		while (i--)
-		{
-			system.EnqueueFiber([&]() {
-				std::this_thread::sleep_for(std::chrono::milliseconds(1));
-				doneCount++;
-
-				return HeartFiberStatus::Complete;
-			});
-		}
-
-		while (doneCount < JobCount)
-			std::this_thread::yield();
-
-		// With so many long-running jobs, we expect every thread to have been hit
-		EXPECT_EQ(threadIds.size(), settings.threadCount);
-
-		// And they should've all finished
-		EXPECT_EQ(doneCount, JobCount);
-
-		system.Shutdown();
-	}
-
-	EXPECT_EQ(allocator.m_allocatedCount, 0);
-}
